refactor(shipping): Phase 9 reshape Shipping.Tests into Features/
- Move single-slice tests to Features/<Slice>/, namespaces renamed to
  Shipping.Tests.Features.<Slice>. Split multi-source files per slice:
  ShipmentTransitionEndpointsTests -> Pick/Pack/CancelShipment slices;
  ShipmentTerminalTransitionTests -> Deliver/Fail/ReturnShipment;
  ShipmentDispatchEndpointsTests -> GetCarrierQuotes + DispatchShipment;
  ShipmentOwnershipTests -> GetShipmentsByOrder + GetShipmentById.
  SeedShipmentAsync helpers duplicated per split file (rule-of-three).
- Relocate cross-cutting tests: InternalOutboxEndpointsTests -> Infrastructure/
  Outbox/ mirror; Carriers/* -> Infrastructure/Carriers/ mirror;
  HealthChecksTests to project root. Observability/ShippingMetricsTests
  stays put (out of scope for this phase).
- Workaround namespace/type clash in Cancel/CreateShipmentCommand slices:
  test namespace Shipping.Tests.Features.CancelShipmentCommand shadows the
  shared lib type; alias to SagaCancelShipmentCommand /
  SagaCreateShipmentCommand at use sites.
- Delete emptied Api/ and Carriers/ test folders. Build green, 107/107
  Shipping.Tests pass (7 Phase-10 skips unchanged).
