+++ shipping-microservice/Shipping.Tests/Infrastructure/Carriers/CarrierPollingServiceTests.cs
@@ -5,7 +5,7 @@ using Shipping.Service.Contracts.Integration;
 using Shipping.Service.Domain;
 using Shipping.Service.Infrastructure.Carriers;
 
-namespace Shipping.Tests.Carriers;
+namespace Shipping.Tests.Infrastructure.Carriers;
 
 public class CarrierPollingServiceTests : IntegrationTestBase
 {
